+++ ocp_tessellate/defaults.py
@@ -79,7 +79,7 @@ class Defaults:
         - ambient_intensity   Intensity of ambient light (default=0.75)
         - direct_intensity    Intensity of direct lights (default=0.15)
         - show_parent:        Show the parent for edges, faces and vertices objects (default=False)
-        - show_locals:        In build123d show local part/sketch/line in addition to the relocated object (default=True)
+        - show_locals:        In build123d show local part/sheet/sketch/line in addition to the relocated object (default=True)
         - show_bbox:          Show bounding box (default=False)
         - tools:              Show the viewer tools like the object tree (default=True)
         - glass:              Show the viewer in glass mode, i.e (CAD navigation as transparent overlay (default=False)